Disable sqlx pool timeouts to avoid losing our DB
It seems that once every connection to an in-memory SQLite DB is closed
the database is dropped. When we use a connection pool it will
eventually close all connections.

We could make sure that the schema and intiial data is created on every
connection if necessary, but using an in-memory DB is supposed to be a
temporary thing for development, so for now just disable connection
timeouts, as per <https://github.com/launchbadge/sqlx/issues/1647#issuecomment-1020677837>